MAINT: Move rss teaser code to own module
Its a block after all
